requirements.txt
Django>=4.0,<5.0
psycopg2>=2.9
gunicorn
gdal==3.4.0
numpy>=1.21.0
djangorestframework
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
fonttools==4.57.0
fpdf==1.7.2
h11==0.14.0
idna==3.10
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pycryptodome==3.22.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.6.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
pydantic>=1.10
flake8>=5.0
black==25.1.0